% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_qsm_segments.R
\name{assign_qsm_segments}
\alias{assign_qsm_segments}
\title{assign_segments}
\usage{
assign_qsm_segments(cylinders)
}
\arguments{
\item{cylinders}{either a string specifying a text file or a data.frame with the cylinder text output file from TreeQSM (need to set inputs.savetxt = 1 in create_input.m of TreeQSM)}

\item{output_file}{name of the xml file to be written}

\item{subdiv}{integer number of subdivisions to be used in tube construction}

\item{data_fields}{vector of column names from input_file that should be written as cylinder object data in the xml file}

\item{branch_file}{optional branch text output file from TreeQSM (need to set inputs.savetxt = 1 in create_input.m of TreeQSM)}
}
\description{
assign_segments
}
\note{
if data_fields is not provided, all columns of the TreeQSM output (cylinder_file and branch_file if provided) are written as object data

if branch_file is provided "branch_" is appended to the beginning of each of that file's column names before joining to the cylinder file. If data_fields is provided, append "branch_" when referencing these column names.
}
