% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beam_diameter_at_distance.R
\name{get_beam_diameter_at_distance}
\alias{get_beam_diameter_at_distance}
\title{get the diameter of a LiDAR beam at distance}
\usage{
get_beam_diameter_at_distance(
  distance,
  beam_diameter_at_exit = 7,
  beam_divergence = 0.3
)
}
\arguments{
\item{distance}{distance between scanner and target (m)}

\item{beam_diameter_at_exit}{the diameter of the LiDAR beam as it exits the scanner (mm)}

\item{beam_divergence}{the divergence of the beam (mrad)}
}
\description{
get the diameter of a LiDAR beam at distance
}
\note{
output is the beam diameter at the target in mm

default values of beam_diameter_at_exit and beam_divergence are for the VZ-1000 scanner
}
