% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_height_range.R
\name{get_target_height_range}
\alias{get_target_height_range}
\title{get the min and max target z coordinates scanned given the scanner distance, height, and scan pattern}
\usage{
get_target_height_range(
  distance,
  scanner_height,
  theta_min = 30,
  theta_max = 130
)
}
\arguments{
\item{distance}{horizontal distance between scanner and target (m)}

\item{scanner_height}{of the scanner (m)}

\item{theta_min}{the minimum zenith angle (deg.) at which the scanner sends out a beam (defaults to a value of 30 for the VZ-1000 scanner)}

\item{theta_max}{the maximum zenith angle (deg.) at which the scanner sends out a beam (defaults to a value of 130 for the VZ-1000 scanner)}
}
\description{
get the min and max target z coordinates scanned given the scanner distance, height, and scan pattern
}
