% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_facet_labels.R
\name{overall_facet_labels}
\alias{overall_facet_labels}
\title{add overall facet labels to a ggplot}
\usage{
overall_facet_labels(
  p,
  labelR = "",
  labelT = "",
  where = "top",
  input = "ggplot"
)
}
\arguments{
\item{p}{base plot that overall facet labels should be added to}

\item{labelR}{the name of the overall variable for the facets on the right side of the plot}

\item{labelT}{the name of the overall variable for the facets on the top side of the plot}

\item{where}{either "top", "right", or "both" - depends on directions of facets}

\item{input}{if "ggplot" a conversion to grob is done, otherwise it is assumed that p is a grob already}
}
\description{
add overall facet labels to a ggplot
}
