% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsm_cylinders_to_Helios_tubes_xml.R
\name{qsm_cylinders_to_Helios_tubes_xml}
\alias{qsm_cylinders_to_Helios_tubes_xml}
\title{qsm_tubes_to_Helios_xml}
\usage{
qsm_cylinders_to_Helios_tubes_xml(
  cylinder_file,
  output_file,
  subdiv = 5,
  data_fields
)
}
\arguments{
\item{cylinder_file}{cylinder text output file from TreeQSM (need to set inputs.savetxt = 1 in create_input.m of TreeQSM)}

\item{output_file}{name of the xml file to be written}

\item{subdiv}{integer number of subdivisions to be used in tube object construction}

\item{data_fields}{vector of column names from input_file that should be written as cylinder object data in the xml file}
}
\description{
qsm_tubes_to_Helios_xml
}
