% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_location_box_buffer.R
\name{scan_location_box_buffer}
\alias{scan_location_box_buffer}
\title{get the location of a bounding box that encompasses all passed in scan locations + a buffer distance}
\usage{
scan_location_box_buffer(x, y, z, buffer)
}
\arguments{
\item{x}{x coordinates of scan locations}

\item{y}{coordinates of scan locations}

\item{z}{coordinates of scan locations}

\item{buffer}{desired buffer distance}
}
\description{
get the location of a bounding box that encompasses all passed in scan locations + a buffer distance
}
