% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ERK.R
\name{theme_ERK}
\alias{theme_ERK}
\title{custom theme}
\usage{
theme_ERK(facets = FALSE)
}
\arguments{
\item{facets}{if TRUE, decreases spacings between facet panels for use with facet_rep_grid from lemon package}
}
\description{
custom theme
}
\note{
this is a custom theme
}
